+++ data/CommonData.lua
@@ -3,8 +3,11 @@ local _, PER = ...
 PER.RACE_ORDER = {"RACE_1", "RACE_2", "RACE_3", "RACE_4", "RACE_5", "RACE_6"}
 PER.DIFFICULTY_ORDER = {"NORMAL", "ADVANCED", "REVERSE", "CHALLENGE", "CHALLENGE_REVERSE", "STORM_GRYPHON"}
 
+PER.COLOR_SILVER = "ffC0C0C0"
+PER.COLOR_BRONZE = "ffCD7F32"
+
 --------------------
---- Data sorting ---
+--- Data Sorting ---
 --------------------
 
 local raceDataTable = PER.RACE_DATA